+++ src/bika/lims/browser/worksheet/ajax.py
@@ -18,80 +18,11 @@
 # Copyright 2018-2021 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
-import json
-from operator import itemgetter
-
 import plone
 import plone.protect
-from bika.lims import FieldEditAnalysisResult
-from bika.lims.api.security import check_permission
 from Products.Archetypes.config import REFERENCE_CATALOG
 from Products.CMFCore.utils import getToolByName
 from Products.Five.browser import BrowserView
-from senaite.core.p3compat import cmp
-
-
-class AttachAnalyses(BrowserView):
-    """ In attachment add form,
-        the analyses dropdown combo uses this as source.
-        Form is handled by the worksheet ManageResults code
-    """
-    def __init__(self, context, request):
-        self.context = context
-        self.request = request
-
-    def __call__(self):
-        plone.protect.CheckAuthenticator(self.request)
-        searchTerm = 'searchTerm' in self.request and self.request['searchTerm'].lower() or ''
-        page = self.request['page']
-        nr_rows = self.request['rows']
-        sord = self.request['sord']
-        sidx = self.request['sidx']
-        analysis_to_slot = {}
-        for s in self.context.getLayout():
-            analysis_to_slot[s['analysis_uid']] = int(s['position'])
-        analyses = list(self.context.getAnalyses(full_objects=True))
-        # Duplicates belong to the worksheet, so we must add them individually
-        for i in self.context.objectValues():
-            if i.portal_type == 'DuplicateAnalysis':
-                analyses.append(i)
-        rows = []
-        for analysis in analyses:
-            # Skip analyses that cannot be edited (attachment not permitted)
-            if not check_permission(FieldEditAnalysisResult, analysis):
-                continue
-            parent = analysis.getParentTitle()
-            rows.append({'analysis_uid': analysis.UID(),
-                         'slot': analysis_to_slot[analysis.UID()],
-                         'service': analysis.Title(),
-                         'parent': parent,
-                         'type': analysis.portal_type})
-
-        # if there's a searchTerm supplied, restrict rows to those
-        # who contain at least one field that starts with the chars from
-        # searchTerm.
-        if searchTerm:
-            orig_rows = rows
-            rows = []
-            for row in orig_rows:
-                matches = [v for v in row.values()
-                           if str(v).lower().startswith(searchTerm)]
-                if matches:
-                    rows.append(row)
-
-        rows = sorted(rows, cmp=lambda x, y: cmp(x, y), key=itemgetter(sidx and sidx or 'slot'))
-        if sord == 'desc':
-            rows.reverse()
-        pages = len(rows) / int(nr_rows)
-        pages += divmod(len(rows), int(nr_rows))[1] and 1 or 0
-        start = (int(page)-1) * int(nr_rows)
-        end = int(page) * int(nr_rows)
-        ret = {'page': page,
-               'total': pages,
-               'records': len(rows),
-               'rows': rows[start:end]}
-
-        return json.dumps(ret)
 
 
 class SetAnalyst(BrowserView):