+++ src/bika/lims/interfaces/__init__.py
@@ -606,6 +606,17 @@ class IIdServerVariables(Interface):
         """
 
 
+class IIdServerTypeID(Interface):
+    """Marker interface for type id resolution for ID Server
+    """
+
+    def get_type_id(self, **kw):
+        """Returns the type id for the context passed in the constructor, that
+        is used for custom ID formatting, regardless of the real portal type of
+        the context. Return None if no type id can be resolved by this adapter
+        """
+
+
 class IReferenceWidgetVocabulary(Interface):
     """Return values for reference widgets in AR contexts
     """
@@ -839,6 +850,8 @@ class IProxyField(Interface):
     """A field that proxies transparently to the field of another object.
     Mainly needed for AnalysisRequest fields that are actually stored on the
     Sample.
+
+    TODO: Remove after the field manager has been removed from `senaite.jsonapi`
     """